app/script/requirements.txt
accelerate==1.6.0
datasets==3.5.1
fastapi==0.115.12
python-dotenv==1.1.0
tensorboard==2.19.0
tokenizers==0.21.1
torch==2.7.0
torchaudio==2.7.0
torchvision==0.22.0
transformers==4.51.3
uvicorn==0.34.2
